refactored code to have the feed section and implemented working sorting
