+++ src/Components/UI/Centered.js
@@ -1,0 +1,7 @@
+import classes from "./Centered.module.css";
+
+function Centered(props) {
+  return <div className={classes.main}>{props.children}</div>;
+}
+
+export default Centered;
