+++ src/Components/UI/Centered.module.css
@@ -1,0 +1,7 @@
+.main {
+  display: flex;
+  margin: 1.5rem auto;
+  width: inherit;
+  justify-content: center;
+  align-items: center;
+}
